Update Google Sheets components from CJS to ESM
