+++ components/google_sheets/actions/create-worksheet/create-worksheet.mjs
@@ -1,6 +1,6 @@
-const googleSheets = require("../../google_sheets.app");
+import googleSheets from "../../google_sheets.app.js";
 
-module.exports = {
+export default {
   key: "google_sheets-create-worksheet",
   name: "Create Worksheet",
   description: "Create a blank worksheet with a title",
