+++ components/google_sheets/actions/delete-row/delete-row.mjs
@@ -1,9 +1,9 @@
-const googleSheets = require("../../google_sheets.app");
+import googleSheets from "../../google_sheets.app.js";
 
-module.exports = {
-  key: "google_sheets-delete-worksheet",
-  name: "Delete Worksheet",
-  description: "Delete a specific worksheet",
+export default {
+  key: "google_sheets-delete-row",
+  name: "Delete Row",
+  description: "Deletes a specific row in a spreadsheet",
   version: "0.0.1",
   type: "action",
   props: {
@@ -35,6 +35,12 @@ module.exports = {
       type: "string",
       label: "Worksheet",
     },
+    row: {
+      propDefinition: [
+        googleSheets,
+        "row",
+      ],
+    },
   },
   async run() {
     const request = {
@@ -42,8 +48,13 @@ module.exports = {
       requestBody: {
         requests: [
           {
-            deleteSheet: {
-              sheetId: this.worksheetId,
+            deleteDimension: {
+              range: {
+                "sheetId": this.worksheetId,
+                "dimension": "ROWS",
+                "startIndex": this.row - 1,
+                "endIndex": this.row,
+              },
             },
           },
         ],
