--- components/google_sheets/actions/get-values/get-values.js
@@ -1,45 +1,0 @@
-const googleSheets = require("../../google_sheets.app");
-
-module.exports = {
-  key: "google_sheets-get-values",
-  name: "Get Values",
-  description: "Get all values from a sheet.",
-  version: "0.0.13",
-  type: "action",
-  props: {
-    googleSheets,
-    drive: {
-      propDefinition: [
-        googleSheets,
-        "watchedDrive",
-      ],
-      description: "",
-    },
-    sheetId: {
-      propDefinition: [
-        googleSheets,
-        "sheetID",
-        (c) => ({
-          driveId: googleSheets.methods.getDriveId(c.drive),
-        }),
-      ],
-    },
-    sheetName: {
-      propDefinition: [
-        googleSheets,
-        "sheetName",
-        (c) => ({
-          sheetId: c.sheetId,
-        }),
-      ],
-    },
-  },
-  async run() {
-    const sheets = this.googleSheets.sheets();
-
-    return (await sheets.spreadsheets.values.get({
-      spreadsheetId: this.sheetId,
-      range: `${this.sheetName}`,
-    })).data.values;
-  },
-};
