+++ components/google_sheets/actions/get-values/get-values.mjs
@@ -1,10 +1,10 @@
-const googleSheets = require("../../google_sheets.app");
+import googleSheets from "../../google_sheets.app.js";
 
-module.exports = {
-  key: "google_sheets-get-values-in-range",
-  name: "Get Values in Range",
-  description: "Get values from a range of cells using A1 notation.",
-  version: "0.0.5",
+export default {
+  key: "google_sheets-get-values",
+  name: "Get Values",
+  description: "Get all values from a sheet.",
+  version: "0.0.13",
   type: "action",
   props: {
     googleSheets,
@@ -33,19 +33,13 @@ module.exports = {
         }),
       ],
     },
-    range: {
-      propDefinition: [
-        googleSheets,
-        "range",
-      ],
-    },
   },
   async run() {
     const sheets = this.googleSheets.sheets();
 
     return (await sheets.spreadsheets.values.get({
       spreadsheetId: this.sheetId,
-      range: `${this.sheetName}!${this.range}`,
+      range: `${this.sheetName}`,
     })).data.values;
   },
 };
